+++ CHANGES.rst
@@ -5,6 +5,7 @@ UNRELEASED
 ----------
 - Replace flyback_pixels property with calculate_flyback_pixels method.
 - Change 2D ramp (tableau) to use relative control values, like 1D ramp.
+- Fix regression in multiple shift EELS acquisition.
 
 0.20.7 (2022-07-26)
 -------------------